correct xpms

+++ proj/src/model/button/button.c
@@ -6,7 +6,7 @@ bool (is_cursor_over_button)(button_t button, position_t mouse_position){
   return is_between_x && is_between_y;
 }
 
-void (change_button_colors)(button_t* button, uint8_t new_background_color, uint8_t new_text_color){
+void (change_button_colors)(button_t* button, int new_background_color, int new_text_color){
   button->background_color = new_background_color;
   button->text_color = new_text_color;
 }
