+++ proj/src/model/button/button.h
@@ -9,8 +9,8 @@ struct button {
     uint16_t y;
     uint16_t width;
     uint16_t height;
-    uint8_t background_color;
-    uint8_t text_color;
+    int background_color;
+    int text_color;
     char *text;
     void (*onClick)(struct button *);
 };
@@ -18,7 +18,7 @@ typedef struct button button_t;
 
 bool (is_cursor_over_button)(button_t button, position_t mouse_position);
 int (get_hovered_button)(button_t *buttons, int num_buttons, position_t mouse_position);
-void (change_button_colors)(button_t* button, uint8_t new_background_color, uint8_t new_text_color);
+void (change_button_colors)(button_t* button, int new_background_color, int new_text_color);
 
 int (process_buttons_clicks)(button_t *button, int num_buttons, drawing_position_t before, drawing_position_t next, int *button_clicked);
